delete not working tests
